Remove other colormodels from "Brightness/Contrast" video plugin
Leave only 16 bit RGB and YUV with alpha
Implement function reset_plugin
Use SPDX-License tag

+++ plugins/brightness/brightnesswindow.C
@@ -1,23 +1,7 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
 
-/*
- * CINELERRA
- * Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
- * 
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- * 
- */
+// This file is a part of Cinelerra-CVE
+// Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
 
 #include "bctitle.h"
 #include "brightnesswindow.h"
@@ -90,16 +74,17 @@ int BrightnessSlider::handle_event()
 
 char* BrightnessSlider::get_caption()
 {
-	float fraction;
+	double fraction;
+
 	if(is_brightness)
 	{
-		fraction = *output / 100;
+		fraction = *output / 100.0;
 	}
 	else
 	{
 		fraction = (*output < 0) ? 
-			(*output + 100) / 100 : 
-			(*output + 25) / 25;
+			(*output + 100.0) / 100.0 :
+			(*output + 25.0) / 25.0;
 	}
 	sprintf(string, "%0.4f", fraction);
 	return string;
